Added handlePlan to torben class
